PriceMonitor/backendapi/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn
wheel
marshmallow
